moved two methods
